Initial runs controller
