+++ core/src/main/scala/io/paradoxical/aetr/core/db/Storage.scala
@@ -60,7 +60,7 @@ class Storage @Inject()(stepDb: StepDb) {
    * @return
    */
   def loadRun(root: RootId): Run = {
-    stepDb.getRun(root).waitForResult()
+    stepDb.getRunTree(root).waitForResult()
   }
 
   /**
