+++ core/src/main/scala/io/paradoxical/aetr/core/server/AetrServer.scala
@@ -5,7 +5,7 @@ import com.twitter.finagle.http.{Request, Response}
 import com.twitter.finatra.http.HttpServer
 import com.twitter.finatra.http.filters.{CommonFilters, LoggingMDCFilter, TraceIdMDCFilter}
 import com.twitter.finatra.http.routing.HttpRouter
-import io.paradoxical.aetr.core.server.controllers.{PingController, StepsController}
+import io.paradoxical.aetr.core.server.controllers.{PingController, RunsController, StepsController}
 import io.paradoxical.aetr.core.server.serialization.JsonModule
 import io.paradoxical.finatra.swagger.{ApiDocumentationConfig, SwaggerDocs}
 
@@ -34,5 +34,6 @@ class AetrServer(override val modules: Seq[Module]) extends HttpServer with Swag
     router
       .add[PingController]
       .add[StepsController]
+      .add[RunsController]
   }
 }
